requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
mysql-connector-python==8.0.26
python-dotenv==0.19.0
pydantic==1.8.2